psvita-robot-console/tools/requirements.txt
flask==3.0.0
pillow==10.1.0
numpy==1.26.0
werkzeug==3.0.0
